+++ packages/common/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @powersync/common
 
+## 1.34.0
+
+### Minor Changes
+
+- ab33799: Add experimental support for raw tables, giving you full control over the table structure to sync into.
+  While PowerSync manages tables as JSON views by default, raw tables have to be created by the application
+  developer.
+
+  For more information about raw tables, see [the documentation](https://docs.powersync.com/usage/use-case-examples/raw-tables).
+
+- 810c6ad: Propagate logger from PowerSyncDatabase to streaming sync and remote implementations, and tweak some log messages.
+
+### Patch Changes
+
+- a1aa18c: Fix sync stream delays during CRUD upload.
+- 9fb898d: [Internal] Removed shared mutex implementation of `readLock` and `writeLock`.
+
+## 1.33.2
+
+### Patch Changes
+
+- 9b2bde3: Fix compilation error due to broken import in `SyncProgress.d.ts`
+
+## 1.33.1
+
+### Patch Changes
+
+- ffe3095: Improve websocket keepalive logic to reduce keepalive errors.
+- 36d8f28: Remove unused compaction logic.
+- 53236a8: Rust client: Properly upload CRUD entries made while offline.
+- b7255b7: Rust sync client: Fix `connect()` resolving before a connection is made.
+- 70a9cf5: Added CommonJs output for common package. Ensuring default export entry is last so that require statements don't use .mjs output instead of .cjs.
+- d1b7fcb: Rust sync client: Fix reported `lastSyncedAt` values in sync status.
+
 ## 1.33.0
 
 ### Minor Changes